+++ examples/04_choices.c
@@ -31,9 +31,7 @@ clags_choice_t choice_values[] = {
 // Choices are case-sensitive by default, with `.case_insensitive=true` you can disable that behaviour
 clags_choices_t choices = clags_choices(choice_values, .case_insensitive=true);
 // Create variable to hold the pointer to the chosen `clags_choice_t`
-// This sets the default value to the first `clags_choice_t` in `choices`
-// This expands to `clags_choice_t *choice = &choice_values[0];`
-clags_choice_t *choice = clags_choice_value(choice_values, 0);
+clags_choice_t *choice = NULL;
 
 clags_arg_t args[] = {
     clags_positional(&input, "input_file", "the input file"),
@@ -41,7 +39,7 @@ clags_arg_t args[] = {
     clags_option('o', "output", &output, "FILE", "the output file"),
 
     // With `.choices` you set the `clags_choices_t` wrapper as the verifier for choice parsing
-    clags_option('a', "algorithm", &choice, "ALG", "the algorithm to use", .value_type=Clags_Choice, .choices=&choices),
+    clags_option('a', "algorithm", &choice, "ALG", "the algorithm to use", .value_type=Clags_Choice, .choices=&choices, .default_input="LIFO"),
     clags_option('q', "quality", &quality, "LEVEL", "the sample quality", .value_type=Clags_UInt8),
 
     clags_flag('w', NULL, &warnings, "print warnings"),